[refactor] Move LLVMType into /core and fix naming
