+++ src/main/kotlin/dev/supergrecko/kllvm/core/LLVMValue.kt
@@ -1,0 +1,5 @@
+package dev.supergrecko.kllvm.core
+
+import org.bytedeco.llvm.LLVM.LLVMValueRef
+
+public class LLVMValue internal constructor(internal val llvmValue: LLVMValueRef)
