+++ src/main/kotlin/dev/supergrecko/kllvm/core/types/LLVMFunctionType.kt
@@ -1,5 +1,6 @@
-package dev.supergrecko.kllvm.core.type
+package dev.supergrecko.kllvm.core.types
 
+import dev.supergrecko.kllvm.core.LLVMType
 import dev.supergrecko.kllvm.utils.iterateIntoType
 import dev.supergrecko.kllvm.utils.toBoolean
 import org.bytedeco.javacpp.PointerPointer
