+++ src/main/kotlin/dev/supergrecko/kllvm/core/types/LLVMIntegerType.kt
@@ -1,11 +1,12 @@
-package dev.supergrecko.kllvm.core.type
+package dev.supergrecko.kllvm.core.types
 
+import dev.supergrecko.kllvm.core.LLVMType
 import org.bytedeco.llvm.LLVM.LLVMTypeRef
 import org.bytedeco.llvm.global.LLVM
 
 public class LLVMIntegerType internal constructor(llvmType: LLVMTypeRef) : LLVMType(llvmType) {
     /**
-     * Returns the amount of bits this integer type can hold
+     * Returns the amount of bits this integer types can hold
      */
     public fun typeWidth(): Int {
         return LLVM.LLVMGetIntTypeWidth(llvmType)
