+++ src/test/kotlin/dev/supergrecko/kllvm/core/types/LLVMStructureTypeTest.kt
@@ -1,6 +1,6 @@
-package dev.supergrecko.kllvm.core.type
+package dev.supergrecko.kllvm.core.types
 
-import dev.supergrecko.kllvm.core.LLVMContext
+import dev.supergrecko.kllvm.core.LLVMType
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
